etl/requirements.txt
elasticsearch==7.13.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pydantic==1.10.2
